Add a benchmark `use`ing the whole `std`
In our benchmark suite for a while (#12025) we had a benchmark that
invoked the `nu-std::load_standard_library` function. But as we
prioritized the startup speed the default behavior is to just import the
prelude and make the modules available for later parsing.

With this added benchmarks the entire library is loaded via `use std`.
This gives us a parsing benchmark with a somewhat diverse set of
constructs that get used. This should be helpful to inspect general
tredns when we make changes to the parser or the parts of the engines
relevant during parse time.