Change the output of `format bits` to big endian instead of native endian (#16435)
Currently the command ouputs in native endian. That is unlike any other
tool that I've used that shows bits (windows calculator comes to mind).
Moreover, if you paste the output of `format bits` as a number you won't
get the same output:

```nushell
~> 258 | format bits
00000010 00000001
~> 0b00000010_00000001
513
```

In order to get a roundtrip, you need to reverse the bytes:

```nushell
~> 0b00000001_00000010
258
~> 258 | format bits | split words | str join | $"0b($in)" | into int
513
~> 258 | format bits | split words | reverse | str join | $"0b($in)" | into int # Added `reverse`
258
```

The goal of this PR is to get rid of this weirdness (and as a bonus get
a platform-independent behavior, without depending on what the native
endianness is).

## Release notes summary - What our users need to know
### Change the output of `format bits` to big endian instead of native
endian

This is what most people expect when formatting numbers and other
primtives in binary. However, until now, if you were in a little endian
system, you would get:
```nushell
~> 258 | format bits
00000010 00000001
```
If you copied and pasted that as a number, you would get a surprising
result:
```nushell
~> 0b00000010_00000001
513
```
Now, `format bits` always formats in big endian:
```nushell
~> 258 | format bits
00000001 00000010
```