Fix `native-tls` feature (#16518)
While building the docs I realized that `nu-command` complained that it
could not find `ureq::TlsConnector`. I investigated a bit and found out
that #16078 broke it (I merged that, lol). It also broke the
`native-tls` feature we have, but it did not come up in our CI as this
would massively increase the CI time. This PR fixes the `native-tls`
feature.

So `cargo build --no-default-features --features native-tls` works
again.