add `$nu.is-lsp` to help with printing in lsp mode (#16635)
This PR adds `$nu.is-lsp` primarily so we can detect when nushell is in
LSP mode programmatically. Where this is immediately beneficial is with
`print` statements. By default `print` statements go to stdout. If this
happens in LSP mode, the LSP will crash because it's looking for
json-rpc on stdout. With the change to the `print` command in this PR,
if nushell is in LSP mode, it only prints to stderr regardless of
whether you specified --stderr or not.

Another idea I had was to make this a global setting where everything in
nushell would print to stderr if in LSP mode but how to do that wasn't
immediately obvious to me.

## Release notes summary - What our users need to know
When nushell is launched with the `--lsp` flag, nushell will set
`$nu.is-lsp` to true so that users can programmatically know when
nushell is in LSP mode.

When in LSP mode, the `print` command will only print to stderr. This is
done to prevent the LSP from misbehaving when it sees text that isn't
JSON-RPC in stdout, which is being listened to by the LSP.

## Tasks after submitting
N/A