Expand alias when sending command to external completion (#16640)
An attempt to fix #8483.

When an alias is used, `flatten_expression` returns a
`FlatShape::External(span)` which contains the expanded alias. Use that
to get the expanded alias and send it to external completion instead of
the alias itself.

```
alias glo = git log --oneline

// Without patch
glo test<TAB>   =>   external completer called with ["glo" "log" "--oneline" "test"}

// With patch
glo test<TAB>   =>   external completer called with ["git" "log" "--oneline" "test"}
```

## Release notes summary - What our users need to know
### Aliases are now expanded before being sent to external completers

Before this change, when an external completer was invoked for an alias,
the first argument sent to the completer would be the alias itself
instead of the aliased command.

For example, with an alias defined as `alias gl = git log`, typing `gl
branch_name` and pressing `TAB` would send the arguments `gl`, `log`,
`branch_name` to the external completer.

After this change, the alias is expanded and the arguments sent to the
external completer will be `git`, `log`, `branch_name`.


## Tasks after submitting
<!-- Remove any tasks which aren't relevant for your PR, or add your own
-->
- [ ] Update the
[documentation](https://github.com/nushell/nushell.github.io)