Fix ir eval error on branched block redirection (Fixes #16582)
*Description*

Attempt to fix issue [#16582](https://github.com/nushell/nushell/issues/16582)
by avoiding duplicated `write` and `close` calls for an `out>`/`err>`
redirected file for `try-catch` and `if-else` blocks.

Similarly to [this PR](https://github.com/nushell/nushell/pull/15617)
a function to identify if a code block is being handeled, and if so skip
the invocation of `finish_redirection` so that no additional writes are
added. (The blocks perform redirections on their own and this
then avoids the duplication).

*User-Facing Changes*

The following commands will no longer raise ir-eval errors:

 - `if true { echo "hey" } out> /dev/null`
 - `if false { echo "hey" } else { echo "ho" } out> /dev/null`
 - `try { 1/0 } catch { echo "hi" } out> /dev/null`

*Tests*

Added 8 tests.