Set up ast-grep with some custom lints (#16674)
ast-grep is a structural search and replace tool and linter based on tree-sitter.

Running `ast-grep scan` will give you a list of diagnostics based on the
rules defined in `ast-grep/rules`.

For editor integration, `ast-grep` supports LSP and has a vscode plugin.

Because it works with tree-sitter, it's possible to define lints for nu
scripts too.