Improve error messages for invalid binary strings (#16688)
Fixes #9566. Based on the decision from this PR #16678, instead of
dealing with out-of-range octal strings, throwing a specific error is
better.

Let me know if these error messages are too detailed and prefer just
having a single "out-of-range error".

## Release notes summary - What our users need to know

Previously, a generic shell error is thrown for invalid binary strings.
Now, an improve error message is thrown for invalid binary, hexadecimal,
and octal strings.
```nushell
> 0b[121]
# => Error: nu::parser::invalid_binary_string
# => 
# => × Invalid binary string.
# =>   ╭─[entry #1:1:1]
# => 1 │ 0b[121]
# =>   · ───┬───
# =>   ·    ╰── invalid binary string
# =>   ╰────
# =>  help: binary strings may contain only 0 or 1.
```
## Tasks after submitting

- N/A