Theming fix (#16722)
## Release notes summary - What our users need to know
### boolean coloring in light theme
When using `std/config light-theme`, booleans inside structures are now
colored in dark cyan instead of white, which was very hard to read in
combination with a light terminal background.

### changed fallback for unknown color names
If some named color in `$env.config.color_config` is misspelled or
otherwise unknown to nu, the corresponding element is now colored using
the terminal's default (as in `ansi reset`) instead of white.