Add configuration for columnar menu traversal direction (#16724)
Make columnar menu tab traversal direction configurable. The
configurable traversal direction feature was added to `reedline` with
https://github.com/nushell/reedline/pull/951. This PR makes that
configuration option available in `nushell`.

## Release notes summary - What our users need to know

The feature makes it possible to choose the direction that items are
placed in the columnar menu. With the default value of `"horizontal"`,
items are ordered left-to-right in each row, with later items appearing
in rows below earlier items.
```
1   2   3   4
5   6   7   8
9   10
```
With the value set to `"vertical"`, items are ordered top-to-bottom in
each column, with later items appearing in columns to the right of
earlier items.
```
1   4   7   10
2   5   8
3   6   9
```

The configuration option is currently called "tab_traversal" with a
default value of "horizontal".
A specific revision of reedline is pointed to in cargo.toml since the
required changes are not available in any released version of reedline.



## Tasks after submitting
<!-- Remove any tasks which aren't relevant for your PR, or add your own
-->
- [ ] Update the
[documentation](https://github.com/nushell/nushell.github.io)

---------

Co-authored-by: Darren Schroeder <343840+fdncred@users.noreply.github.com>