update nushell to use coreutils v0.2.2 crates (#16737)
This is a follow up to https://github.com/nushell/nushell/pull/15896 and
https://github.com/nushell/nushell/pull/15932 and updates nushell to use
the coreutils crates v0.2.2. We should try to stay on the latest
coreutils version as much as possible.

Coreutils has reverted the main problem we had which was the need to
have a `PROJECT_NAME_FOR_VERSION_STRING` environment variable.
For reference,

https://github.com/uutils/coreutils/blob/5287738e9c408cf665e5a87c4df03bdc1685bb8f/docs/src/release-notes/0.2.0.md?plain=1#L528
https://github.com/uutils/coreutils/pull/8360

## Release notes summary - What our users need to know
N/A

## Tasks after submitting
N/A