feat: add `not-starts-with` and `not-ends-with` operators (#16852)
<!--
Thank you for improving Nushell!
Please, read our contributing guide:
https://github.com/nushell/nushell/blob/main/CONTRIBUTING.md
-->

Since no one seems opposed to it
https://discord.com/channels/601130461678272522/615329862395101194/1426617232074608650

## Release notes summary - What our users need to know
<!--
This section will be included as part of our release notes. See the
contributing guide for more details.
Please include only details relevant for users here. Motivation and
technical details can be added above or below this section.

You may leave this section blank until your PR is finalized. Ask a core
team member if you need help filling this section.
-->
Added `not-starts-with` and `not-ends-with` operators to be on par with
other comparison operators,

## Tasks after submitting
<!-- Remove any tasks which aren't relevant for your PR, or add your own
-->
- [x] Add tests
- [x] Update the
[documentation](https://github.com/nushell/nushell.github.io).
https://github.com/nushell/nushell.github.io/pull/2054
- [x] Update the tree-sitter grammar.
https://github.com/nushell/tree-sitter-nu/pull/227