add `explore regex` command (#16910)
The PR adds a new command to the `explore` family, `explore regex`. 

## Release notes summary - What our users need to know
The new `explore regex` command is meant to provide a TUI for exploring
regular expressions and visualizing the results in real time. You can
pipe text into the command and that text becomes the "test string" that
the regular expressions run on. You can also start it up without piping
anything into it and manually type the "test string" and then the regex.
You can use the tab key to switch between panes. The arrow keys work to
navigate as we as the page up and page down keys, home, and end.

![explore-regex](https://github.com/user-attachments/assets/150722a8-d398-478a-a689-30290d3caf9c)


## Tasks after submitting
<!-- Remove any tasks which aren't relevant for your PR, or add your own
-->
- [ ] Update the
[documentation](https://github.com/nushell/nushell.github.io)