str expand: two dots outside of curly braces should not be considered as range operator (#16924)
## Release notes summary - What our users need to know
This fixes #16739
No need to escape dots, literally dots (`.`) outside of curly braces
while using bracoxide. I haven't thought that would be useful in some
cases. It was not a bug, but it was a faulty design choice of mine. Some
users didn't understand the logic behind it as well as I dont either.
I'm questioning why did i leave it like that. Definitely gotta improve
that library.

## Tasks after submitting
<!-- Remove any tasks which aren't relevant for your PR, or add your own
-->
- [ ] Update the
[documentation](https://github.com/nushell/nushell.github.io)
I'm not sure about documentation. Should I? It is really a small change
but a life-saver to some.

Thanks, sincerely.