test: Set HOME for test_config_path tests (#16914)
This PR sets the `HOME` environment in the test playground if the outer
process has it set. Without it, if `$env.HOME` is set in the process
running tests, the playground process won't have this variable, and so,
inside the playground, Nushell will use the password file to find the
home directory on Linux. This causes three of the config path tests to
fail on Linux when the two sources for home directory disagree.

This was found by @KaiSforza, who found in
[#16848](https://github.com/nushell/nushell/pull/16848/files#diff-d20879fe15880267dc80df70a17d396186ab2fa5887e58ed57e3aa2bbf4572b2R42-R45)
that in Nix's sandbox environment, three tests
(`test_default_config_path`, `test_xdg_config_bad`,
`test_xdg_config_empty`) were failing due to the `HOME` env var in the
environment not matching the home directory set in the password file.

I confirmed that with this change, building the `default.nix` provided
in the linked PR worked without having to skip the three previously
failing tests.

## Release notes summary - What our users need to know

N/A