Allow setting span in custom/extern completers (#15888)
# Description

Closes https://github.com/nushell/nushell/issues/15863. This PR allows
custom and extern completers to set the `span` field for the
`Suggestion`s they provide.

The motivation for this is mainly making it easier to make completers
for Zoxide.

# Release notes summary

Custom completers and external completers will be able to choose which
part of the buffer to replace. This is useful if you also want to
replace one of the previous arguments to a command, or if you only want
to replace part of an expression.

# Tests + Formatting

Didn't add any tests, but did manually try the following:
```
> def comp [context, pos] { [{value: mwahahaha, description: bar, span: {start: 0, end: $pos}}] }
> def foo [arg: string@comp] {}
```
After creating those definitions, `foo <TAB>` will replace the command
line with `mwahahaha`.

# After Submitting

Needs to be mentioned in the completer docs.