Add `--strict` flag to `first` and `last` command, to raise error on empty list. (#17066)
As title, this pr adds `--strict` flag to `first` and `last` command, so
they will raise error on empty list. This is a follow up to #17054

## Release notes summary - What our users need to know
By default, `first` and `last` will return `null` on empty list, this pr
is going to add `--strict` to make it returns on error.

```
> [] | first --strict
Error: nu::shell::access_beyond_end

  × Row number too large (empty content).
   ╭─[entry #1:1:6]
 1 │ [] | first --strict
   ·      ──┬──
   ·        ╰── index too large (empty content)
   ╰────
```

## Tasks after submitting
NaN