Add `error_lines` and `error_style = nested` settings
With the new inner `error make` options, the `.show_related_errors_as_nested()`
setting makes a lot of sense, I but did not want to enable it by default. This
adds a new error mode that uses the nested error style.

Also added a setting to change the number of context lines shown for the labels.
Sometimes this can let labels get squashed into one block of code, making it
easier to read.