fix: throw errors on overflow in duration integer multiplication (#17063)
Partially fixes #17062.

This is a small change which adds checks for all potentially overflowing
duration/integer operations, in addition to the sum check added in
#16594. This doesn't include floating point operations, which might
still overflow.

I also took the liberty to move the shared code into a function. It does
mean that the span message will no longer have the operation name. If
that's not acceptable, I can refactor it to use the operator name.

## Release notes summary - What our users need to know

- Multiplying duration by integers now throws an error on overflow