Fix up the env.rs tests (#17085)
Just uses `ends-with` and a prefix to skip the control characters at the
start of the strings

This fixes the issues that came up in #17038 for being run in terminals.
This doesn't fix the underlying issue, however it does allow tests to be
run in certain terminals and still pass.

This also fixes a problem with some of the tests that load `std/log`. If
`$env.NU_LOG_FORMAT` is set in the parent shell, then these tests will
incorrectly fail. Clearing the env out before running makes these work.

<!--
Thank you for improving Nushell!
Please, read our contributing guide:
https://github.com/nushell/nushell/blob/main/CONTRIBUTING.md
-->

## Release notes summary - What our users need to know
<!--
This section will be included as part of our release notes. See the
contributing guide for more details.
Please include only details relevant for users here. Motivation and
technical details can be added above or below this section.

You may leave this section blank until your PR is finalized. Ask a core
team member if you need help filling this section.
-->

## Tasks after submitting
<!-- Remove any tasks which aren't relevant for your PR, or add your own
-->
- [ ] Update the
[documentation](https://github.com/nushell/nushell.github.io)