fix: Make custom completer spans relative to command expr (#17060)
Changes spans provided by custom/external completers to be relative to
the context/line provided as input to the completer. Previously, they
were relative to the entire command line, which would cause problems if
the partial command call were part of a subexpression.

For example, if the user typed `foo | bar ba<TAB>` and the completer
provided the span `{start: 0, end: 6}`, the `foo | ` bit incorrectly
gets replaced. With this PR, `bar ba` would be replaced.

To clarify, this PR does not change how `$pos` is passed to custom
completers. It only changes how spans in suggestions provided by the
completer are interpreted.

Fixes #17046

## Release notes summary - What our users need to know
<!--
This section will be included as part of our release notes. See the
contributing guide for more details.
Please include only details relevant for users here. Motivation and
technical details can be added above or below this section.

You may leave this section blank until your PR is finalized. Ask a core
team member if you need help filling this section.
-->

Note to release note writers: while this changes behavior, I don't think
this is behavior anyone was counting on, and so isn't really a breaking
change. I don't think it's necessary to include this in the release
notes at all, but I'll add a summary anyway and leave it to your
discretion.

[#15888](https://github.com/nushell/nushell/pull/15888) gave custom and
external completers the ability to provide a span for each suggestion to
choose which part of the typed text would be replaced. However, these
spans were treated as being relative to the entire buffer, causing
problems if the command call being completed were a subexpression. With
this release, they will be treated as relative to just the expression
passed to the completer.

## Tasks after submitting

- [ ] Update the
[documentation](https://www.nushell.sh/book/custom_completions.html) to
talk about custom spans