Add quick reference panel to `explore regex` command (#17187)
This PR adds a quick reference panel to `explore regex` with the F1
keybinding and dismisses it with the ESC keybinding. If you click on any
of the items in the quick reference panel, it puts them in the Regex
Pattern edit box. I also refactored a bit to make it eassier to
maintain.

<img width="3234" height="2070" alt="image"
src="https://github.com/user-attachments/assets/502a6566-715f-4ce0-b787-6012fc5b9942"
/>


## Release notes summary - What our users need to know
Add a quick reference panel to `explore regex` with the F1 keybinding
and dismiss it with the ESC keybinding.

## Tasks after submitting
N/A