Stream explore nu output (#17208)
This is an attempt to get explore's `:nu` command to stream the output
back to explore without blocking the UI. You can test with this command:
```nushell
:nu seq 1 100 | each { |x| sleep 100ms; $x }
```
The `:nu:` command is pretty heavily modified to use channels to
communicate now while spawning a background thread for processing.

This ups the complexit quite a bit so I won't be offended if we decide
_not_ to land this command.

closes https://github.com/nushell/nushell/issues/17206

## Release notes summary - What our users need to know
Within the `explore` command the `:nu` command will now stream content
by spawning the command issues without blocking.

## Tasks after submitting
N/A