Value::List and Value::Range can be interrupted when piping to other commands. (#17245)
Fixes: #14862
This issue is happened when creating Iterator from `List/Range` value,
the signal is set to `Signals::empty`, which is unwanted behavior
because `Singal::empty` can'e be interruped.

To fix it, we need to make `List/Range` carries a signal from
`engine_state.signals`. I did this by:
1. introducing `signals` fields to `Value::List/Value::Range`
2. before running a command, manually `inject_signals` to the value.

To be honest the implementation is not really good, not sure if there is
a better way to do this :-(

## Release notes summary - What our users need to know
### Infinite sequences cant be terminated when piped to other command
```
> `1..inf | to md`  # then press ctrl-c
^CError: nu::shell::error

  × Operation interrupted
   ╭─[entry #1:1:1]
 1 │ 1..inf | to md
   · ───────┬──────
   ·        ╰── This operation was interrupted
   ╰────
```